Clique + Ion unit tests with mocked classes. Rearchitected and enhanced entire suite of contracts.
