+++ package.json
@@ -16,6 +16,7 @@
     "solc": "^0.4.23",
     "solhint": "^1.1.10",
     "solidity-coverage": "^0.5.0 ",
+    "truffle-assertions": "^0.6.3",
     "web3-eth-accounts": "^1.0.0-beta.34"
   },
   "devDependencies": {
@@ -37,9 +38,9 @@
   "scripts": {
     "testrpc": "ganache-cli --account='0xe176c157b5ae6413726c23094bb82198eb283030409624965231606ec0fbe65b,1000000000000000000000000000000000000000' --account='0xd18bc3878eb28192238d92ae085cdb9438527e36faa92484dea2e3baa047b083,1000000000000000000000000000000000000000' --port 8545 --gasLimit 0xFFFFFFFFFFF --gasPrice 0 --networkId 1234 --deterministic grape crisp enroll avoid satoshi picnic term dice gown grocery situate depart",
     "compile": "truffle compile",
-    "deploy": "truffle deploy --network development",
-    "depclique": "truffle deploy --network clique",
-    "test": "truffle test --network clique",
+    "deploy": "truffle deploy",
+    "test": "truffle test",
+    "debug": "truffle debug",
     "coverage": "solidity-coverage",
     "lint": "solhint contracts/**/*.sol",
     "genbin": "solc --overwrite --bin ./contracts/*.sol -o abi",
