+++ truffle.js
@@ -3,6 +3,7 @@ module.exports = {
     development: {
       host: "localhost",
       port: 8545,
+      gas: 0xFFFFFFFFFFF,
       network_id: "*"
     },
     clique: {
